script/python/summarizer/requirements.txt
nltk==3.0.0
networkx>=1.9
psycopg2>=2.5.4
pyyaml>=3.11
